Fix to bb function and adjustment to raster convention.
Depending on the angle, the order of min and max for both x and y can
change in the bounding box function. We now sort to ensure the
expected ordering.

The raster function avoids a final flip by flipping the vertical index
on assignments. There has to be a flip somewhere, and this is the most
direct way to do it (assign to array rows in correct order by the
origin=upper convention)