No more flips.
This follows the principle of least surprise. Flipping the result of
certain calculations caused a disagreement between the order of rows
in the result and the order of points specified in the axis. Now
result and grid remain in agreement. Orientation for plotting is
handled in the imshow method in grid.